==> 05-results.Rmd <==
# Results

```{r}
library(ggridges)
library(lubridate)
library(tidyverse)
library(gridExtra)
```

```{r}
commit <- read_csv("resources/data/commits.csv") %>% filter(year(date) >= 2014)
issue <- read_csv("resources/data/issues.csv")  %>% filter(year(created_at) >= 2014)
pull <- read_csv("resources/data/pull_requests.csv")  %>% filter(year(created_at) >= 2014)
issue_and_pull_request <- bind_rows(issue %>% mutate(category = "issue"), pull %>% mutate(category = "pull request"))
```

## Users prefer weekdays or weekends?

It is natural to guess that users will tend to do most of their word (submit commits, create issues and pull requests) during weekdays. And it is the case for commits of both packages and pull requests of plotly, where there are a steep drop in numbers. But surprisingly, there are still quite a lot of pull requests (drops for about 50%) are submitted on weekends for ggplot2, and the number of issues created on weekends for both packages are also not that small.


```{r, fig.height=8, fig.width=8}
# function to make bar plot, number versus wday
count_per_wday <- function(tibble, category){
  tibble %>%
    select(repo, date) %>%
    mutate(created_wday = wday(date,label=T)) %>%
    count(repo, created_wday, name = "count") %>%
    mutate(category = category)
}
bind_rows(
  count_per_wday(commit, "commit"),
  count_per_wday(issue %>% rename(date = created_at), "issue"),
  count_per_wday(pull %>% rename(date = created_at), "pull_request")
) %>%
  ggplot(aes(x = created_wday, y = count)) +
  geom_col() +
  facet_grid(category~repo, scales = "free_y") +
  labs(x = "created at (week)", 
       title = "ggplot2's pull requests are still pretty high during weekends",
       subtitle = str_wrap("number of commits/issues/pull requests created over weekdays"),
       caption = "Note: Y axis' scale is different for different rows")
```

## Slow down during pandemic?

We can see that almost all the numbers are decreasing over the past few years except for ggplot2's issues and plotly's pull requests. Although it's tempting to think that the pandemic may actually slow down both packages' development, yet there is no causality here and there may be many factors behind such decreases. Also we can see that the number of issues created over the past few years hardly decreases for ggplot2. But strangely, plotly has very few pull requests over the last four years (almost zero pull request on 2017 and 2020)


```{r, fig.height=8}
# function to make line chart, number versus year
count_per_year <- function(tibble, category){
  tibble %>%
    select(repo, date) %>%
    mutate(year = year(date)-2000) %>%
    count(repo, year, name = "count") %>%
    mutate(category = category)
}
bind_rows(
  count_per_year(commit, "commit"),
  count_per_year(issue %>% rename(date = created_at), "issue"),
  count_per_year(pull %>% rename(date = created_at), "pull_request")
) %>%
  ggplot(aes(x = year, y = count)) + 
    geom_line() + geom_point() +
    facet_grid(category~repo, scales = "free_y") +
    scale_x_continuous(breaks=14:21) +
  labs(x = "created at (year)", 
       title = "There is indeed a steep drop over last three years",
       subtitle = str_wrap("number of commits/issues/oull requests over 2014-2021"),
       caption = "Note: Y axis' scale is different for different rows")

```

## Who are the most active users over years?

Let's break down the number of issues and pull requests to the level of user's roles. Due to the similar meaning in the definition of "COLLABORATOR" and "MEMBER" in these two packages, we combine these two categories into one, "DEVELOPER". (author association's definition can be found [here](https://docs.github.com/en/graphql/reference/enums#commentauthorassociation)) As expected, issues are mainly created by "NONE" users and pull requests are mainly created by "COLLABORATOR / MEMBER" users. But there is a clear downward trend for "DEVELOPER" users of both packages, especially in ggplot2's pull request graph.


```{r, fig.height=8, fig.width=10}
# function to make multi-line plot, number for different author_association versus year
multi.line_plot <- function(tibble){
  tibble %>%
    select(repo, author_association, created_at, category) %>%
    mutate(created_year = year(created_at) - 2000) %>%
    mutate(author_association = 
             ifelse(author_association %in% c("MEMBER", "COLLABORATOR"), 
                    "DEVELOPER", author_association)) %>%
    count(repo, category, author_association, created_year, name = "number") %>%
    ggplot(aes(x = created_year, y = number, color = author_association)) +
    geom_line() + geom_point() +
    facet_grid(category~repo, scale = "free_y") + 
    scale_x_continuous(breaks = 14:21)
}
# issue of different author association per year
multi.line_plot(issue_and_pull_request) +
  labs(x = "created at (year)", y = "number", color = "author association",
       title = "None has most issues while developer has most pull requests as expected",
       subtitle = str_wrap("breakdown of number of issues/pull requests over different author assocation"),
       caption = "Note: Y axis' scale is different for different rows")

```

## Do different roles get different number of comments?

We can see that the overall distribution skew to the right a lot regardless of packages, categories (issue or pull request), or author association, especially for plotly. And in general, ggplot2 's users get more comments than plotly's. But it is hard to believe that 25% of pull requests of plotly's developer and 50% of issues created by plotly's developer receives 0 comments. 

```{r}
box_plot <- function(tibble){
  tibble %>%
    select(repo, author_association, comments, category) %>%
    filter(comments < 15) %>%
    mutate(author_association = 
             ifelse(author_association %in% c("MEMBER", "COLLABORATOR"), 
                    "DEVELOPER", author_association)) %>%
    ggplot(aes(x = comments, y = repo)) +
    geom_boxplot() +
    facet_grid(category~author_association)
}
box_plot(issue_and_pull_request) +
  labs(title = "plotly's developer tend to receives very few comments",
       subtitle = str_wrap("ggplot2 vs plotly, number of comments for different author association")) 
```

## Are most of issues/pull requests still unsolved?

Issue and pull request can have two status, open and closed, which indicates if the issue has been solved or if the pull request has been accepted/rejected. We can see that both packages have very few open pull requests but plotly has far more open issues than ggplot both in relative and absolute scale.


```{r}
issue_and_pull_request %>% 
  count(repo, category, state, name = "count") %>% 
  ggplot(aes(x = state, y = count, fill = repo, label = count)) + 
  geom_col(position = "dodge") + 
  geom_text(position = position_dodge(width = 1), vjust = -0.5) +
  facet_wrap(~category) +
  labs(title = "Plotly has more issues that are still open",
       subtitle = "ggplot2 v.s. plotly, number of open/close issues and pull requests")
```

## Are open cases percentage increasing?

Let's first dive into the open issues. We can see that the percentage of "open" issues (issues whose status are still open) increases for plotly, especially for the past three years. Although we would expect some increases since earlier created issues are given more time to solve, yet plotly's increases is substantial (from 30%+ in 2019 to 50%+ in 2020 and to 70%+ in 2021). When compared to ggplot, plotly's percentage of both issues and pull requests are much higher even though plotly has fewer issue and pull requests than ggplot in absolute terms.


```{r}
issue_and_pull_request %>%
  select(created_at, closed_at, repo, author_association, state, repo, category) %>%
  mutate(created.year = (format(created_at, "%y")) %>% as.factor) %>%
  count(repo, category, created.year, state, name = "count") %>%
  group_by(repo, category, created.year) %>% 
  mutate(perc = count / sum(count)) %>%
  filter(state == "open") %>%
  ggplot(aes(x = created.year, y = perc, fill = repo)) +
  geom_col(position = "dodge") +
  scale_y_continuous(labels = scales::percent) +
  facet_wrap(~category) +
  labs(x = "created at (year)", y = "percentage" ,
       title = "% of issues whose states are still open increases for Plotly",
       subtitle = str_wrap("ggplot2 v.s. plotly, % of issues and pull requests created each year whose status are still open over 2014-2021"),
       caption = "Note: percentage is calculated over total issues/pull requests created within each year")
```

## How long to solve issues/pull requests?

Next, let's move our attention to closed issues and pull requests. We calculate issues and pull requests' solved time (which is defined as close time - create time) and divided the time into periods of 30 days (or 1 month). And we can see most of issues/pull requests were closed within one month but plotly seems to have larger percentage of issues that were open for more than one year.


```{r, fig.width=8}
time.period <- c(0, seq(30, 360, 30), 2000)
time.labels <- c(sapply(1:12, function(x){paste("<", x, " month", sep="")}), ">1 year")

issue_and_pull_request_perc <- issue_and_pull_request %>%
  select(created_at, closed_at, repo, author_association, state, repo, category) %>%
  mutate(created.year = (format(created_at, "%y")) %>% as.factor,
         solve.time = interval(created_at, closed_at) / days(1),
         solve.period = cut(solve.time, breaks=time.period, labels=time.labels)) 

issue_and_pull_request_perc %>%
  count(category, repo, solve.period, name = "count") %>%
  group_by(category, repo) %>% 
  mutate(perc = count / sum(count)) %>%
  filter(! is.na(solve.period)) %>%
  ggplot(aes(x = solve.period, y = perc, fill = repo)) +
  geom_col(position="dodge") +
  scale_y_continuous(labels = scales::percent) +
  facet_wrap(~category, ncol = 1) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5)) +
  labs(x = "time period", y ="percentage", 
       title = "Most of issues/pull requests were closed within one month",
       subtitle = "% of issues/pull requests whose solved time falls into each periods of 1 month",
       caption = "Note: percentage is calculated over both open and closed issues/pull requests")

```

## Do different years exhibit different patterns to solve issue/pull request?

Breaking down the percentage over 2014-2021, we can see that both packages seems to having a trend of taking longer to "close" their issues. But the pattern is very different for pull requests, the percentage of pull requests closes within one month increases for the past several years. The contrast is most spark for plotly in 2021, only about 25% of issues are closed within two months, but over 80% of pull requests are closed within one month.


```{r, fig.height=6}
issue_and_pull_request_perc %>%
  count(category, repo, created.year, solve.period, name = "count") %>%
  group_by(category, repo, created.year) %>% 
  mutate(perc = count / sum(count)) %>%
  filter(! is.na(solve.period)) %>%
  ggplot(aes(x = created.year, y = solve.period, fill = perc)) +
  geom_tile() +
  scale_fill_gradient(low="white", high="blue") +
  facet_grid(category~repo) + 
  labs(x = "created at (year)", y = "time period", fill = "percentage", 
       title = "Both packages seems to take longer to solve issues",
       subtitle = "a detailed breakdown of percentages over 2014-2021",
       caption = "Note: percentage is calculated over both open and closed issues/pull requests within each year")
```
